% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeller_master.R
\name{labeller_master}
\alias{labeller_master}
\title{Re-label facets when using ggplot2's \code{\link[ggplot2]{facet_grid}}.}
\usage{
labeller_master(label_list)
}
\arguments{
\item{label_list}{A list mapping old labels to new labels.}
}
\value{
A ggplot2 labeller function that can be passed as the \code{labeller}
argument for \code{\link[ggplot2]{facet_grid}}.
}
\description{
\code{labeller_master} returns a function that can be used to re-label factes
when using ggplot2's \code{\link[ggplot2]{facet_grid}}.
}
\details{
\code{labeller_master} returns a function that can be used to re-label factes
when using ggplot2's \code{\link[ggplot2]{facet_grid}}. The input list should
map the default facet labels to new facet labels. This function will only
work when facetting on 1 variable.
}

