% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nzapply.R
\name{nzapply}
\alias{nzapply}
\title{Apply a vectorized function to all nonzero elements of a vector.}
\usage{
nzapply(x, FUN, base = 0, ...)
}
\arguments{
\item{x}{A vector.}

\item{FUN}{A vectorized function.}

\item{base}{What to return if all elements of x are zero.}

\item{...}{Other arguments to pass to \code{FUN}}
}
\value{
Returns the value of applying \code{FUN} to nonzero elements of
  \code{x}
}
\description{
\code{nzapply} applies a vectorized function to all nonzero elements of a
vector.
}
\details{
Source:

\url{http://stackoverflow.com/questions/12583569/how-can-i-calculate-the-means-of-rows-while-excluding-the-zero-values-from-rows}
}
\examples{
nzapply(c(0, 1, 2, 3, 4), min)
nzapply(c(0, 0, 0, 0), max, base = NA)
}

