% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_group_panel.R
\name{plot_group_panel}
\alias{plot_group_panel}
\title{Plot boxplots of log10 relative abundance, facetted by 1 feature}
\usage{
plot_group_panel(metadata, otu_tab_matrix, group, base_group = "Control",
  cutoff = min(20, nrow(otu_tab_matrix)), alpha = 0.05, notch = F,
  title = "", labeller = "label_value")
}
\arguments{
\item{metadata}{Data frame of metadata, like the one returned by readPCL.}

\item{otu_tab_matrix}{matrix of relative abundance, like the one returned by
readPCL. Each row should be an OTU, and each column should be a sample.}

\item{group}{The grouping variable in the metadata.}

\item{base_group}{The reference level for the grouping variable
\code{group}. When performing hypothesis tests, all other levels of the
grouping variable \code{group} are compared to this one.}

\item{cutoff}{Only display the top \code{cutoff} bugs in terms of median
relative abundance.}

\item{alpha}{alpha level to conduct each hypothesis test.}

\item{notch}{Whether to have notched boxplots.}

\item{title}{Title for the graph.}

\item{labeller}{labeller function to pass to ggplot2's
\code{\link[ggplot2]{facet_grid}}. The default uses the values of the labels.}
}
\value{
ggplot2 plotting object. This function does not automatically display
the plot, so you need to call \code{\link{print}} on its output to display
the plot.
}
\description{
\code{plot_group_panel} plots boxplots of log10 relative abundances, sorted
by their median and facetted by 1 feature. In each plot, the top
\code{cutoff} OTUs are plotted, as well as a line connecting the medians from
the \code{base_group}. Any increases or decreases in the other groups,
relative to the base group, are colored. The main inputs are a
\code{metadata} \code{\link{data.frame}}, and a matrix for the OTU table. The
metadata should be in \emph{long} format, with one row corresponding to one
sample. The OTU table should have the OTUs as the rows, and the samples as
the columns.
}
\details{
\code{plot_group_panel} plots boxplots of log10 relative abundances, sorted
by their median and facetted by 1 feature. In each plot, the top
\code{cutoff} OTUs are plotted, as well as a line connecting the medians from
the \code{base_group}. Any increases or decreases in the other groups,
relative to the base group, are colored. The main inputs are a
\code{metadata} \code{\link{data.frame}}, and a matrix for the OTU table. The
metadata should be in \emph{long} format, with one row corresponding to one
sample. The OTU table should have the OTUs as the rows, and the samples as
the columns.

The main inputs are a \code{metadata} \code{\link{data.frame}}, and a matrix
for the OTU table. The metadata should be in \emph{long} format, with one row
corresponding to one sample. The OTU table should have the OTUs as the rows,
and the samples as the columns.

The coloring is determined by the results of \code{\link{hypothesis_test}},
which currently uses a Mann-Witney U test as implemented in
\code{\link{wilcox.test}}.
}

