% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_bug_names.R
\name{process_bug_names}
\alias{process_bug_names}
\title{Pretty print taxonomy name}
\usage{
process_bug_names(name, sep = "; ")
}
\arguments{
\item{name}{Unprocessed, full taxonomy name.}

\item{sep}{String separating different levels of taxonomic names}
}
\value{
Return value will be a string that includes classification levels
with consecutive alphabetical characters as a sub-level of the main level <4
}
\description{
\code{process_bug_names} cleans up full taxonomy names to display the most
relevant information
}
\details{
The classification levels in the name should be prefixed by k__, p__, etc.
for kingdom, phylum, etc (like greengenes profiles).

Tries to return genus + species, if species is present. If genus is not
present (e.g. contains less than 4 consecutive alphabetical characters),
tries to return the next highest level of classification. If that is not
present, keeps trying.
}

