% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_bug_names.R
\name{process_silva_name}
\alias{process_silva_name}
\title{Pretty print taxonomy name (silva-like)}
\usage{
process_silva_name(name, sep = ";")
}
\arguments{
\item{name}{Unprocessed, full taxonomy name.}

\item{sep}{String separating different levels of taxonomic names}
}
\value{
Return value will be a string that includes classification levels
with consecutive alphabetical characters as a sub-level of the main level <4
}
\description{
\code{process_silva_name} cleans up full taxonomy names to display the most
relevant information.
}

