% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readPCL.R
\name{readPCL}
\alias{readPCL}
\title{Read numeric data plus metadata from a .pcl file}
\usage{
readPCL(filename, number.pheno.rows = NA, ...)
}
\arguments{
\item{filename}{Fully specified path to a .pcl file}

\item{number.pheno.rows}{Number of rows of metadata. If
\code{number.pheno.rows} is NA, this function will figure out the number of
rows of phenotypes, under the assumption that 5 consecutive rows of numeric
values signifies the start of the numeric data at the startn of those
consecutive rows. If \code{number.pheno.rows} is specified, it tells the
program how many rows at the start of the file (not including the header) are
phenotype data.}

\item{...}{Additional arguments passed on to \code{\link{read.table}}.}
}
\value{
Returns a list object containing two elements:
\enumerate{
    \item phenodat, a dataframe containing the potentially non-numeric
      phenotype data.
    \item  numericdat, a matrix containing the numeric expression data.
}
}
\description{
\code{readPCL} reads numeric data plus metadata from a .pcl file. It assumes
that the first row is a header. It can attempt to automatically determine
which rows are metadata and which are phenotype data, or the number of
phenotype rows can be specified explicitly.
}

